Attempt to Add Numbering of Chart
This code change is an attemp to add numbers in the corner of every table
cell. Unfortuneately the changes have made a bit of a mess of the code.

There is a need to refactor everything.

This is a switch from the originial idea of having a new set of rows with
number appear bewlow the sybmols.