Creates Middle Chart and End Chart Numbers Work
Much rewriting of code.
Numbers can now be faded in at the appropriate time.

All numbers are working and in their correct location.

CSS is in better shape.

Random symbol switching is completely redone in a much better way.

Currently in progress is getting the secret symbols in the correct spots.