Finished Chart Switching
The full basic concept is finsished although the code needs refacturing. Also
the instructions as to what to do needs to be added.

Instructions:
Pick a two digit number. Add both digits together. Subtract the
result from the original number. Then click anywhere on the page.

A chart will be created and numbered. Find the result of your calculation
on the chart and memorize the matching symbol.

Click anywhere on the page again and concentrate on the symbol you found
in the chart. That symbol will fill the chart.