Refactored mindCrypt.js to module pattern
Everything is now in a single module called mindCrypt. mindCrypt.init() is
used to start the app, and is bound to jQuery.ready()

Many of the parts of the app are now structured into internal functions of
the module.

It could still use some work to make it cleaner and more intuitive to read
but it's much better.

Also added .jshintrc file to help with the refactoring.