Animated appearance of Psycic Connection Icon
This code (mostly CSS tweaking) get's everything working correctly so the
Psycic Connection Icon can move in and out of view. Currently the code
demostrates the animation.

Also CSS was tweked so messages will float to the bottom near the chart
withouth the chart moving up and down during the animation.