responsive design done
All responsive functionality should be working. The chart shrinks
according to the width of the screen. When it reaches the width of a
vertical phone the chart switchest to a full width and to a two by
three square set up.

All responsiveness is done through Bootsrap classes except the last
mentioned thing because there is no Bootstrap 3 class out of the box for
widths smaller than 480. Jquery is therefore used to switch out the
Bootstrap classes when this widths is reached.