keeps global space clean
Moves the pop up message variables to the main mindCrypt module. Everything is
now in the document ready function or the mindCrypt module.

The code is still not organized in any good fashion however at least the
global space is now pretty clean.

Ultimately it would be nice to organize the code into a MVC style.